[refactor] 이미지 최적화 및 productLike 컴포넌트 /commerce에서 공용으로 사용하도록 변경
