+++ dutchiepay/src/app/_components/_commerce/_product/ProductItem.jsx
@@ -5,7 +5,7 @@ import '@/styles/globals.css';
 
 import Image from 'next/image';
 import Link from 'next/link';
-import ProductLike from './ProductLike';
+import ProductLike from '../ProductLike';
 import Rating from '@/app/_components/_rating/Rating';
 
 export default function ProductItem({ item }) {
@@ -31,7 +31,7 @@ export default function ProductItem({ item }) {
             ({item.reviewCount > 999 ? '999+' : item.reviewCount})
           </p>
         </div>
-        <ProductLike isLiked={item.liked} productId={item.buyId} />
+        <ProductLike isLiked={item.liked} productId={item.buyId} size={22} />
       </div>
       <p className="min-h-[48px] mt-[8px] title--multi-line font-medium">
         {item.productName}
