+++ dutchiepay/src/app/_components/_commerce/_productDetail/ProductHeader.jsx
@@ -22,6 +22,7 @@ export default function ProductHeader({ product, productId }) {
             alt={product?.productName}
             fill
             priority
+            sizes="(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw"
           />
         </div>
